Ability to authenticate user before executing CRUD requests

+++ Modules/movie-handler.js
@@ -7,10 +7,10 @@ const movieHandler = {};
 
 // Function - get all movies not equal to user email
 movieHandler.getAllMovies = function(request, expressResponse){
-
+  console.log(request.user);
   // Set property of user email
-  // let queryObject = {email: request.user.email};
-  let queryObject = {}; // deactivate this and re-activate with email once Autho0 established
+  let queryObject = {email: request.user.email};
+  // let queryObject = {}; // deactivate this and re-activate with email once Autho0 established
 
   Movie.find(queryObject) // Find data by user email
     .then(data => {
@@ -43,10 +43,10 @@ movieHandler.createMovie = function(request, expressResponse){
   }
 
   // Add user email property to movie data
-  // let movieDataWithEmailAdded = {...data, email: request.user.email}; ---> Add to Movie.create(movieDatawithEmailAdded)
+  let movieDataWithEmailAdded = {...data, email: request.user.email, userName: request.user.given_name}; //---> Add to Movie.create(movieDatawithEmailAdded)
 
   // Create new movie from data along with user email and send back confirmation if data created successfully
-  Movie.create(data) // --> change to movie.Create(movieDataWithEmailAdded) once Auth0
+  Movie.create(movieDataWithEmailAdded) // --> change to movie.Create(movieDataWithEmailAdded) once Auth0, from Movie.create(data)
     .then(createdMovie => expressResponse.status(201).send(createdMovie))
     .catch(err => {
       // Log error and send a 500 Internal Server Error response
@@ -59,7 +59,7 @@ movieHandler.createMovie = function(request, expressResponse){
 movieHandler.updateMovie = function(request, expressResponse){
   const {id} = request.params;
   const data = request.body;
-  // const userEmail = request.user.email;
+  const userEmail = request.user.email;
 
   // new - returns updated doc instead of old doc
   // overwrite - overwrites doc completely avoiding unwanted properties/side-effects
@@ -70,9 +70,9 @@ movieHandler.updateMovie = function(request, expressResponse){
       }
 
       // Reactivate once using Auth0
-      // if (movie.email !== userEmail) {
-      //   return expressResponse.status(403).send('Unauthorized to update this movie');
-      // }
+      if (movie.email !== userEmail) {
+        return expressResponse.status(403).send('Unauthorized to update this movie');
+      }
 
       // Proceed with update if the user email matches
       return Movie.findByIdAndUpdate(id, data, { new: true, overwrite: true });
@@ -89,7 +89,7 @@ movieHandler.updateMovie = function(request, expressResponse){
 // Function - delete a movie from database
 movieHandler.deleteMovie = function(request, expressResponse){
   const {id} = request.params;
-  // const userEmail = request.user.email;
+  const userEmail = request.user.email;
 
   // First, find movie and check if it belongs to the user by comparing email
   Movie.findById(id)
@@ -99,9 +99,9 @@ movieHandler.deleteMovie = function(request, expressResponse){
       }
 
       // Reactivate - once have Auth0 that passes in email address
-      // if (movie.email !== userEmail) {
-      //   return expressResponse.status(403).send('Unauthorized to delete this Movie');
-      // }
+      if (movie.email !== userEmail) {
+        return expressResponse.status(403).send('Unauthorized to delete this Movie');
+      }
 
       // Proceed with deletion if the user email matches
       return Movie.findByIdAndDelete(id);
